move UsersHelper into model concerns as TimeValidation
